Refactored League related code into a new League directory.
