+++ src/app/league/live-game/page.js
@@ -2,10 +2,10 @@
 
 import React, { useEffect } from "react";
 import { useSearchParams } from "next/navigation";
-import LiveGame from "@/components/LiveGame";
+import LiveGame from "@/components/league/LiveGame";
 import Loading from "@/components/Loading";
-import useProfileData from "../../hooks/useProfileData";
-import NoActiveGameData from "@/components/NoActiveGameData"; // Import your new component
+import useProfileData from "@/app/hooks/league/useProfileData";
+import NoActiveGameData from "@/components/league/NoActiveGameData"; // Import your new component
 
 const LiveGamePage = () => {
   const { liveGameData, isLoading, error, fetchLiveGameData } = useProfileData();
