+++ src/components/league/LiveGame.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import Image from "next/image";
 import Link from "next/link";
-import OutageBanner from "@/components/OutageBanner";
+import OutageBanner from "@/components/league/OutageBanner";
 
 const LiveGame = ({ liveGameData }) => {
 	const [isArena, setIsArena] = useState(false);
