+++ src/components/league/LiveGameBanner.js
