+++ src/components/league/RankedInfo.js
