+++ src/components/league/Tag.js
